requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchaudio>=2.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0
audiomentations>=0.30.0

# Evaluation Metrics
pesq>=0.0.4
pystoi>=0.3.3

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
